+++ sdk/core/azure-core/inc/azure/core/http/policy.hpp
@@ -343,25 +343,14 @@ namespace Azure { namespace Core { namespace Http {
   };
 
   /**
-   * @brief Defines options for getting token.
-   */
-  struct TokenRequestOptions
-  {
-    /**
-     * @brief Authentication scopes.
-     */
-    std::vector<std::string> Scopes;
-  };
-
-  /**
    * @brief Bearer Token authentication policy.
    */
   class BearerTokenAuthenticationPolicy : public HttpPolicy {
   private:
-    std::shared_ptr<TokenCredential const> const m_credential;
-    TokenRequestOptions m_tokenRequestOptions;
+    std::shared_ptr<Credentials::TokenCredential const> const m_credential;
+    Credentials::TokenRequestContext m_tokenRequestContext;
 
-    mutable AccessToken m_accessToken;
+    mutable Credentials::AccessToken m_accessToken;
     mutable std::mutex m_accessTokenMutex;
 
     BearerTokenAuthenticationPolicy(BearerTokenAuthenticationPolicy const&) = delete;
@@ -371,19 +360,19 @@ namespace Azure { namespace Core { namespace Http {
     /**
      * @brief Construct a Bearer Token authentication policy.
      *
-     * @param credential A #Azure::Core::TokenCredential to use with this policy.
-     * @param tokenRequestOptions #Azure::Core::Http::TokenRequestOptions.
+     * @param credential An #Azure::Core::TokenCredential to use with this policy.
+     * @param tokenRequestContext #Azure::Core::Credentials::TokenRequestContext.
      */
     explicit BearerTokenAuthenticationPolicy(
-        std::shared_ptr<TokenCredential const> credential,
-        TokenRequestOptions tokenRequestOptions)
-        : m_credential(std::move(credential)), m_tokenRequestOptions(std::move(tokenRequestOptions))
+        std::shared_ptr<Credentials::TokenCredential const> credential,
+        Credentials::TokenRequestContext tokenRequestContext)
+        : m_credential(std::move(credential)), m_tokenRequestContext(std::move(tokenRequestContext))
     {
     }
 
     std::unique_ptr<HttpPolicy> Clone() const override
     {
-      return std::make_unique<BearerTokenAuthenticationPolicy>(m_credential, m_tokenRequestOptions);
+      return std::make_unique<BearerTokenAuthenticationPolicy>(m_credential, m_tokenRequestContext);
     }
 
     std::unique_ptr<RawResponse> Send(