+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_system_client.cpp
@@ -76,7 +76,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
 
   DataLakeFileSystemClient::DataLakeFileSystemClient(
       const std::string& fileSystemUrl,
-      std::shared_ptr<Core::TokenCredential> credential,
+      std::shared_ptr<Core::Credentials::TokenCredential> credential,
       const DataLakeClientOptions& options)
       : m_fileSystemUrl(fileSystemUrl), m_blobContainerClient(
                                             _detail::GetBlobUrlFromUrl(fileSystemUrl),
@@ -90,11 +90,11 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
             m_fileSystemUrl.GetHost(), options.SecondaryHostForRetryReads));
     perRetryPolicies.emplace_back(std::make_unique<Storage::_detail::StoragePerRetryPolicy>());
     {
-      Azure::Core::Http::TokenRequestOptions tokenOptions;
-      tokenOptions.Scopes.emplace_back(Storage::_detail::StorageScope);
+      Azure::Core::Credentials::TokenRequestContext tokenContext;
+      tokenContext.Scopes.emplace_back(Storage::_detail::StorageScope);
       perRetryPolicies.emplace_back(
           std::make_unique<Azure::Core::Http::BearerTokenAuthenticationPolicy>(
-              credential, tokenOptions));
+              credential, tokenContext));
     }
     {
       Azure::Core::Http::_internal::ValueOptions valueOptions;